requirements.txt
Flask==2.3.3
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
openpyxl==3.1.2
fpdf2==2.7.6
APScheduler==3.10.4
